+++ app/src/main/java/com/techcourse/controller/RegisterController.java
@@ -6,9 +6,11 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
+import nextstep.web.annotation.Controller;
 import nextstep.web.annotation.RequestMapping;
 import nextstep.web.support.RequestMethod;
 
+@Controller
 public class RegisterController {
 
     @RequestMapping(value = "/register", method = RequestMethod.POST)
