--- mvc/src/main/java/nextstep/mvc/controller/asis/ControllerHandlerAdapter.java
@@ -1,22 +1,0 @@
-package nextstep.mvc.controller.asis;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.HandlerAdapter;
-import nextstep.mvc.view.JspView;
-import nextstep.mvc.view.ModelAndView;
-
-public class ControllerHandlerAdapter implements HandlerAdapter {
-
-    @Override
-    public boolean supports(final Object handler) {
-        return handler instanceof Controller;
-    }
-
-    @Override
-    public ModelAndView handle(final HttpServletRequest request, final HttpServletResponse response,
-                               final Object handler) throws Exception {
-        String location = ((Controller) handler).execute(request, response);
-        return new ModelAndView(JspView.from(location));
-    }
-}
