+++ mvc/src/test/java/nextstep/mvc/controller/tobe/HandlerExecutionTest.java
@@ -7,6 +7,7 @@ import static org.mockito.Mockito.when;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.lang.reflect.Method;
+import nextstep.mvc.controller.HandlerExecution;
 import nextstep.mvc.view.ModelAndView;
 import org.junit.jupiter.api.Test;
 import samples.TestController;
